регистрация

+++ bonusprogram/src/main/java/org/itstep/repositories/TransactionRepository.java
@@ -4,6 +4,5 @@ package org.itstep.repositories;
 import org.itstep.domain.Transaction;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TransactionRepository extends CrudRepository<Transaction, Long> {
-
+public interface TransactionRepository extends CrudRepository<Transaction, Integer> {
 }
